# api_app/docs/swagger.yaml
definitions:
...
  api_model.BaseResponseAPIModel:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  api_model.CreateParkingRequestDto:
    properties:
      address:
        type: string
      city_id:
        type: integer
      close_at_hour:
        type: integer
      close_at_minute:
        type: integer
      district_id:
        type: integer
      images:
        items:
          type: string
        type: array
      lat:
        type: number
      lng:
        type: number
      open_at_hour:
        type: integer
      open_at_minute:
        type: integer
      owner_name:
        type: string
      owner_phone:
        type: string
      parking_name:
        type: string
      parking_phone:
        type: string
      parking_types:
        items:
          $ref: '#/definitions/api_model.CreateParkingTypesRequestDto'
        type: array
      ward_id:
        type: integer
    type: object
  api_model.CreateParkingTypesRequestDto:
    properties:
      parking_type:
        type: integer
      price:
        type: number
      total_slot:
        type: integer
    type: object
...
  api_model.GenJWTTokenExternalRequestDto:
    properties:
      token:
        type: string
    type: object
...
  api_model.ImportParkingResponseDto:
    properties:
      total_fail:
        type: integer
      total_success:
        type: integer
    type: object
...
  api_model.RetrieveListParkingActionRolesDto:
    properties:
      can_approve:
        type: boolean
      can_block:
        type: boolean
      can_close:
        type: boolean
      can_denied:
        type: boolean
      can_remove:
        type: boolean
      can_reopen:
        type: boolean
    type: object
  api_model.RetrieveListParkingResponseDto:
    properties:
      parking_name:
        type: string
      parking_phone:
        type: string
      public_id:
        type: string
      roles:
        $ref: '#/definitions/api_model.RetrieveListParkingActionRolesDto'
      status:
        type: integer
      status_display:
        type: string
    type: object
  api_model.UpdateParkingRequestDto:
    properties:
      address:
        type: string
      city_id:
        type: integer
      close_at:
        type: string
      district_id:
        type: integer
      lat:
        type: number
      lng:
        type: number
      open_at:
        type: string
      owner_name:
        type: string
      owner_phone:
        type: string
      parking_name:
        type: string
      parking_phone:
        type: string
      ward_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/internal/v1/:parking_id/sync-es:
    put:
      consumes:
      - application/json
      description: Sync Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Sync Parking
      tags:
      - internal
  /api/internal/v1/{parking_id}/approve:
    put:
      consumes:
      - application/json
      description: Approved Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Approved Parking
      tags:
      - internal
  /api/internal/v1/{parking_id}/close:
    put:
      consumes:
      - application/json
      description: Close Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Close Parking
      tags:
      - internal
  /api/internal/v1/{parking_id}/detail:
    get:
      consumes:
      - application/json
      description: Detail Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Detail Parking
      tags:
      - internal
  /api/internal/v1/{parking_id}/reopen:
    put:
      consumes:
      - application/json
      description: Reopen Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Reopen Parking
      tags:
      - internal
  /api/internal/v1/{parking_id}/update:
    put:
      consumes:
      - application/json
      description: Update parking information
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_id
        in: path
        name: parking_id
        required: true
        type: string
      - description: model
        in: body
        name: model
        schema:
          $ref: '#/definitions/api_model.UpdateParkingRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update parking information
      tags:
      - internal
  /api/internal/v1/add:
    post:
      consumes:
      - application/json
      description: Add Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/api_model.CreateParkingRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Add Parking
      tags:
      - internal
  /api/internal/v1/approve-multi/{parking_ids}:
    put:
      consumes:
      - application/json
      description: Approved Multi Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_ids
        in: path
        name: parking_ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Approved Multi Parking
      tags:
      - internal
...
  /api/internal/v1/generate-external-token:
    post:
      consumes:
      - application/json
      description: Get External JWT Token
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: model
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/api_model.GenJWTTokenExternalRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get External JWT Token
      tags:
      - auth
  /api/internal/v1/import:
    post:
      consumes:
      - application/json
      description: Import Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  $ref: '#/definitions/api_model.ImportParkingResponseDto'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Import Parking
      tags:
      - internal
  /api/internal/v1/retrieve:
    get:
      consumes:
      - application/json
      description: Retrieve paging list parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: page_index
        in: query
        name: page_index
        type: integer
      - description: page_limit
        in: query
        name: page_limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/api_model.RetrieveListParkingResponseDto'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Retrieve paging list parking
      tags:
      - internal
  /api/internal/v1/sync-multi-es/{parking_ids}:
    put:
      consumes:
      - application/json
      description: Sync multi Parking
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: parking_ids
        in: path
        name: parking_ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api_model.BaseResponseAPIModel'
            - properties:
                data:
                  type: boolean
              type: object
      security:
      - ApiKeyAuth: []
      summary: Sync multi Parking
      tags:
      - internal
...
swagger: "2.0"